W1: make core/protocol.ts the single source of truth for the wire contract
Lifts the daemon<->client wire types out of src/daemon/protocol.ts into
core/protocol.ts so the contract has one home that both the daemon and the
clients build against (SessionConfig option a from the plan).

- core/protocol.ts now defines the message shapes and shared domain types:
  error codes, Tx{Slot,PublicState,Intent,Status}, DaemonStatus, decode/tx/
  tx_update/log/error events, and the SessionConfig type (part of the wire
  contract: status carries device; save_config/start_session carry session).
- src/daemon/protocol.ts re-exports those and keeps the server-side DaemonError
  class plus the parse/validate helpers.
- src/daemon/config.ts imports SessionConfig from core and re-exports it; its
  runtime validation/persistence is unchanged.
- tsconfig.json includes core/** (the daemon depends on core for the first
  time); emit is dist/src/... + dist/core/....
- Fix the pre-existing broken start script (node dist/index.js ->
  node dist/src/index.js) to match the actual emit path.

Establishes the source-of-truth half of review finding #6; the client-side
adoption (typed DaemonClient, dropping inline `msg.x as T` casts) lands with
W1 step 3. Verified: build, typecheck, test (55), and dist runtime resolution.
