+++ src/daemon/protocol.ts
@@ -1,107 +1,26 @@
-import type { SessionConfig } from "./config.js";
+// The wire types (message shapes + shared domain types that cross the WebSocket
+// boundary) live in core/protocol.ts as the single source of truth. This module
+// re-exports them and additionally owns the server-side DaemonError class and
+// the parse/validate helpers used only by the daemon.
+export type {
+  CommandId,
+  ErrorCode,
+  TxSlot,
+  TxPublicState,
+  SessionConfig,
+  TxIntent,
+  TxStatus,
+  DaemonStatus,
+  DecodeEvent,
+  TxEvent,
+  TxUpdateEvent,
+  LogEvent,
+  BroadcastEvent,
+  ErrorMessage
+} from "../../core/protocol.js";
+export { errorCodes } from "../../core/protocol.js";
 
-export const errorCodes = [
-  "INVALID_COMMAND",
-  "VALIDATION_FAILED",
-  "CONTROL_REQUIRED",
-  "CONTROL_UNAVAILABLE",
-  "AUTH_FAILED",
-  "CONFIG_REQUIRED",
-  "CONFIG_INVALID",
-  "CONFIG_WRITE_FAILED",
-  "SESSION_ALREADY_ACTIVE",
-  "NO_ACTIVE_SESSION",
-  "SOUND_DEVICE_UNAVAILABLE",
-  "AUDIO_DISCOVERY_FAILED",
-  "CAT_FAILED",
-  "CAT_PORT_UNAVAILABLE",
-  "UDP_BIND_FAILED",
-  "ENGINE_START_FAILED",
-  "PROCESS_CRASHED",
-  "TX_FAILED"
-] as const;
-
-export type ErrorCode = (typeof errorCodes)[number];
-export type CommandId = string | number;
-
-export type TxSlot = "even" | "odd";
-export type TxPublicState = "idle" | "pending" | "active";
-
-export interface TxIntent {
-  af: number;
-  slot: TxSlot;
-  message: string;
-}
-
-export interface TxStatus {
-  state: TxPublicState;
-  af: number | null;
-  slot: TxSlot | null;
-  message: string | null;
-}
-
-export interface DaemonStatus {
-  type: "status";
-  id?: CommandId;
-  session: {
-    active: boolean;
-    mode: "FT8" | null;
-    device: SessionConfig["device"] | null;
-    catConnected: boolean;
-    freq: number | null;
-    ptt: boolean;
-    callsign: string | null;
-    grid: string | null;
-  };
-  tx: TxStatus;
-  control: {
-    held: boolean;
-    byThisClient: boolean;
-  };
-}
-
-export interface DecodeEvent {
-  type: "decode";
-  ts: number;
-  snr: number;
-  dt: number;
-  af: number;
-  mode: "FT8" | "FT4";
-  message: string;
-}
-
-export interface TxEvent {
-  type: "tx";
-  ts: number;
-  af: number;
-  mode: "FT8" | "FT4";
-  message: string;
-}
-
-export interface TxUpdateEvent {
-  type: "tx_update";
-  ts: number;
-  af: number | null;
-  slot: TxSlot | null;
-  message: string | null;
-  state: TxPublicState;
-}
-
-export interface LogEvent {
-  type: "log";
-  level: "info" | "warn" | "error";
-  message: string;
-}
-
-export type BroadcastEvent = DecodeEvent | TxEvent | TxUpdateEvent | LogEvent;
-
-export interface ErrorMessage {
-  id?: CommandId;
-  type: "error";
-  code: ErrorCode;
-  message: string;
-  details?: unknown;
-}
+import type { CommandId, ErrorCode, ErrorMessage, TxIntent } from "../../core/protocol.js";
 
 export class DaemonError extends Error {
   constructor(
